requirements.txt
streamlit
azure-storage-blob
fastapi
uvicorn
python-dotenv
requests
openai
streamlit-chat
